fix time_t printf on x32
On the x32 architecture (x86_64 with 32-bit address space and ints,
but with 64-bit time_t), we see errors like the following:

programs/pluto/connections.c: In function 'add_connection':
programs/pluto/connections.c:1603:74: error: format '%ld' expects argument of type 'long int', but argument 3 has type 'time_t {aka long long int}' [-Werror=format=]
     loglog(RC_LOG_SERIOUS,"IKE lifetime limited to the maximum allowed %lds",
                                                                          ^
programs/pluto/connections.c:1608:76: error: format '%ld' expects argument of type 'long int', but argument 3 has type 'time_t {aka long long int}' [-Werror=format=]
     loglog(RC_LOG_SERIOUS,"IPsec lifetime limited to the maximum allowed %lds",
                                                                            ^

This patch should make the time_t printf work portably by always
casting it to a long long int where it gets printed.

The use of TIME_T_MAX in ttolub also produces a warning:

linux/include/libreswan.h:60:21: error: large integer implicitly truncated to unsigned type [-Werror=overflow]
 #define TIME_T_MAX  ((time_t) ((1ull << (sizeof(time_t) * BITS_PER_BYTE - 1)) - 1))
                     ^
programs/pluto/plutomain.c:942:32: note: in expansion of macro 'TIME_T_MAX'
    ugh = ttoulb(optarg, 0, 10, TIME_T_MAX, &u);
                                ^~~~~~~~~~

This patch also resolves this warning.

+++ programs/pluto/ikev2.c
@@ -1675,15 +1675,15 @@ time_t ikev2_replace_delay(struct state *st, enum event_type *pkind,
 		 */
 		if (IS_IKE_SA_ESTABLISHED(st)) {
 			delay = deltasecs(c->sa_ike_life_seconds);
-			DBG(DBG_LIFECYCLE, DBG_log("ikev2_replace_delay() picked up estalibhsed ike_life:%lu", delay));
+			DBG(DBG_LIFECYCLE, DBG_log("ikev2_replace_delay() picked up estalibhsed ike_life:%lld", (long long int) delay));
 		} else {
 			delay = PLUTO_HALFOPEN_SA_LIFE;
-			DBG(DBG_LIFECYCLE, DBG_log("ikev2_replace_delay() picked up half-open SA ike_life:%lu", delay));
+			DBG(DBG_LIFECYCLE, DBG_log("ikev2_replace_delay() picked up half-open SA ike_life:%lld", (long long int) delay));
 		}
 	} else {
 		/* Delay is what the user said, no negotiation. */
 		delay = deltasecs(c->sa_ipsec_life_seconds);
-		DBG(DBG_LIFECYCLE, DBG_log("ikev2_replace_delay() picked up salifetime=%lu", delay));
+		DBG(DBG_LIFECYCLE, DBG_log("ikev2_replace_delay() picked up salifetime=%lld", (long long int) delay));
 	}
 
 	/* By default, we plan to rekey.
@@ -1900,7 +1900,7 @@ static void success_v2_state_transition(struct msg_digest *md)
 						EVENT_RELEASE_WHACK_DELAY, st);
 				kind = EVENT_SA_REPLACE;
 				delay = ikev2_replace_delay(st, &kind, st->st_original_role);
-				DBG(DBG_LIFECYCLE, DBG_log("ikev2 case EVENT_v2_RETRANSMIT: for %lu seconds", delay));
+				DBG(DBG_LIFECYCLE, DBG_log("ikev2 case EVENT_v2_RETRANSMIT: for %lld seconds", (long long int) delay));
 				event_schedule(kind, delay, st);
 
 			}  else {
@@ -1913,8 +1913,8 @@ static void success_v2_state_transition(struct msg_digest *md)
 			break;
 		case EVENT_SA_REPLACE: /* IKE or Child SA replacement event */
 			delay = ikev2_replace_delay(st, &kind, st->st_original_role);
-			DBG(DBG_LIFECYCLE, DBG_log("ikev2 case EVENT_SA_REPLACE for %s state for %lu seconds",
-				IS_IKE_SA(st) ? "parent" : "child", delay));
+			DBG(DBG_LIFECYCLE, DBG_log("ikev2 case EVENT_SA_REPLACE for %s state for %lld seconds",
+                            IS_IKE_SA(st) ? "parent" : "child", (long long int) delay));
 			delete_event(st);
 			event_schedule(kind, delay, st);
 			break;
